+++ konsave/__main__.py
@@ -1,6 +1,9 @@
 """Konsave entry point."""
 
 import argparse
+import os
+import shutil
+from pkg_resources import resource_filename
 from konsave.funcs import (
     list_profiles,
     save_profile,
@@ -10,7 +13,12 @@ from konsave.funcs import (
     import_profile,
     wipe,
 )
-from konsave.vars import VERSION, list_of_profiles, length_of_lop
+from konsave.vars import (
+    VERSION,
+    CONFIG_FILE,
+    list_of_profiles,
+    length_of_lop,
+)
 
 
 def _get_parser() -> argparse.ArgumentParser:
@@ -90,13 +98,18 @@ def _get_parser() -> argparse.ArgumentParser:
 
 def main():
     """The main function that handles all the arguments and options."""
+
+    if not os.path.exists(CONFIG_FILE):
+        default_config_path = resource_filename("konsave", "conf.yaml")
+        shutil.copy(default_config_path, CONFIG_FILE)
+
     parser = _get_parser()
     args = parser.parse_args()
 
     if args.list:
         list_profiles(list_of_profiles, length_of_lop)
     elif args.save:
-        save_profile(args.save, list_of_profiles, args.force)
+        save_profile(args.save, list_of_profiles, force=args.force)
     elif args.remove:
         remove_profile(args.remove, list_of_profiles, length_of_lop)
     elif args.apply: